% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{apply_sub_file_COV}
\alias{apply_sub_file_COV}
\title{Extracts Covariates for a Subject from a Subject Data File}
\usage{
apply_sub_file_COV(
  tmpcfg,
  cov_found,
  sub_dataset,
  sub_ID_col,
  sub_TIME_col,
  file_ID
)
}
\arguments{
\item{tmpcfg}{ubiquity system object}

\item{cov_found}{list of covariates found in dataset}

\item{sub_dataset}{name of dataset with subject parameters}

\item{sub_ID_col}{name of column in dataset with subject IDs}

\item{sub_TIME_col}{name of column in dataset with simulation time}

\item{file_ID}{subject ID to extract covariates for}
}
\value{
ubiquity system object with the covariates set to those for the current subject
}
\description{
This function is used when stochastic simulations are being performed using
a data file for the subject level information. If the data file contains
covariate information, this function will update the system for each subjects
covariates.
}
\keyword{internal}
