% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{bound_parameters}
\alias{bound_parameters}
\title{Bound Parameters}
\usage{
bound_parameters(pv, cfg)
}
\arguments{
\item{pv}{parameter vector}

\item{cfg}{ubiquity system object}
}
\value{
bounded parameters vector
}
\description{
Takes a vector of parameters and returns the same vector where parameters beyond there bounds are replaced with the bounds
}
\keyword{internal}
