% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{calculate_objective}
\alias{calculate_objective}
\title{Calculates the Value of the Specified Objective Function}
\usage{
calculate_objective(parameters, cfg, estimation = TRUE)
}
\arguments{
\item{parameters}{system parameters}

\item{cfg}{ubiquity system object}

\item{estimation}{boolean variable to indicate if the objective function is being called during parameter estimation}
}
\value{
If estimation is \code{TRUE} it will return the objective function
value, if it is \code{FALSE} it will return a list with an element
\code{value} containing the objective function value and an element named
\code{isgood} that is \cite{TRUE} if the objective function was successful.
}
\description{
For a given set of system parameters the objective function
will be calculated based on defined cohorts and variance models.
}
\keyword{internal}
