% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{calculate_objective_ga}
\alias{calculate_objective_ga}
\title{\code{GA} Wrapper for calculate_objective}
\usage{
calculate_objective_ga(pvect, cfg)
}
\arguments{
\item{pvect}{system parameters}

\item{cfg}{ubiquity system object}
}
\value{
objective function value
}
\description{
Converts the parameter vector to a named list and returns the
negative of the objective (turning the maximization into a minimization)
}
\keyword{internal}
