% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{calculate_objective_pso}
\alias{calculate_objective_pso}
\title{\code{pso} Wrapper for calculate_objective}
\usage{
calculate_objective_pso(pvect, cfg)
}
\arguments{
\item{pvect}{system parameters}

\item{cfg}{ubiquity system object}
}
\keyword{internal}
