% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{calculate_variance}
\alias{calculate_variance}
\title{Calculates the Variance in od_general}
\usage{
calculate_variance(SIMINT_parameters, SIMINT_varstr, SIMINT_odchunk,
  SIMINT_cfg)
}
\arguments{
\item{SIMINT_parameters}{system parameters}

\item{SIMINT_varstr}{string containing variance calculation}

\item{SIMINT_odchunk}{chunk of observation details containing predictions, observations and the time}

\item{SIMINT_cfg}{ubiquity system object}
}
\value{
Variance calculated for a given set of parameters in a model
}
\description{
Takes the variance specified as a string and evaluates it
locally, and returns that value
}
\keyword{internal}
