% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{check_deps}
\alias{check_deps}
\title{Implementation of Matlab \code{toc()} command}
\usage{
check_deps(deps = c(), verbose = FALSE, stop_on_error = FALSE)
}
\arguments{
\item{deps}{list of package dependencies to check (e.g., c("dplyr"))}

\item{verbose}{boolean indiciating if messages should be sent back to the user (default \code{FALSE})}

\item{stop_on_error}{boolean indiciating if messages should be sent back to the user (default \code{FALSE})}
}
\value{
List with the following elements
\itemize{
\item{isgood} TRUE if all the packages were foudn
\item{msgs} Verbose messages
\item{pkgs_missing} Vector of packages that were not found.  
}
}
\description{
Used in conjunction with \code{tic()} to find the elapsed time
when code is executed.
}
\examples{
check_deps(deps=c("dplyr", "ggplot2","bob"), verbose=TRUE)
}
