% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{check_steady_state}
\alias{check_steady_state}
\title{Verify System steady-state}
\usage{
check_steady_state(cfg, som)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{som}{output of \code{\link{run_simulation_ubiquity}}}
}
\description{
Takes the output  \code{\link{run_simulation_ubiquity}} and verifies that the system is running at steady-state by analyzing the timecourse of all of the states in the system
}
