% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Performs parameter estimation}
\usage{
estimate_parameters(cfg)
}
\arguments{
\item{cfg}{ubiquity system object}
}
\value{
list with elements: 
\itemize{
\item \code{estimate} - vector of parameter estimates,
\item \code{raw} - raw output from the underlying optimization routine, 
\item \code{conv} - list of convergence criteria with keys \code{num} (numeric)  \code{text} (text description)
\item \code{obj} - objective function value,  \code{statistics_est} - solution
\item \code{sysup} - 
\item \code{estimate} - 
\item \code{statistics_est} - 
\item \code{report} - 
\item \code{cohorts} - 
\item \code{cohort_view} - 
\item \code{system_file} - 
}
}
\description{
Performs the actual parameter estimation
}
\keyword{internal}
