% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{generate_parameter}
\alias{generate_parameter}
\title{Generates a Parameter Based on \code{<IIV:?>} in the System File}
\usage{
generate_parameter(
  SIMINT_parameters,
  SIMINT_cfg,
  SIMINT_PARAMETER_TV,
  SIMINT_IIV_VALUE,
  SIMINT_equation
)
}
\arguments{
\item{SIMINT_parameters}{parameters vector containing the typical values}

\item{SIMINT_cfg}{ubiquity system object}

\item{SIMINT_equation}{equation relating IIV and typical value to the parameter value with variability}

\item{SIMINT_PARAMETERS_TV}{Typical value of the parameter in question}

\item{SIMINT_PARAMETERS_IIV_VALUE}{sample from mvr distribution}
}
\value{
parameter value with the variability applied
}
\description{
Internal function used to generate parameters based on IIV information
}
\keyword{internal}
