% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{generate_report}
\alias{generate_report}
\title{Generate Text Report with Estimation Results}
\usage{
generate_report(parameters, ss, cfg, conv_desc)
}
\arguments{
\item{parameters}{list of parameter estimates}

\item{ss}{output from solution_statistics}

\item{cfg}{ubiquity system object}

\item{conv_desc}{description of convergence criteria}
}
\value{
List with the following elements: 

\itemize{
  \item \code{report_file} name of report file          
  \item \code{report_file_contents} contents of report file
  \item \code{parameters_all_file} name of CSV file with all parameters 
  \item \code{parameters_est_file} name of CSV file with only the estimates 
}
}
\description{
Internal function used to generate a report of estimation results
}
\keyword{internal}
