% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{generate_subject}
\alias{generate_subject}
\title{Generate Subject}
\usage{
generate_subject(parameters, cfg)
}
\arguments{
\item{parameters}{vector of nominal parameter values}

\item{cfg}{ubiquity system object}
}
\value{
List with a field named \code{parameters} containing a sample representing a subject
}
\description{
Generates subject with variability specified using the \code{<IIV:?>} descriptor
in the system file
}
\keyword{internal}
