% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{GUI_log_entry}
\alias{GUI_log_entry}
\title{Wrapper for system_log_entry Used in Shiny Apps}
\usage{
GUI_log_entry(cfg, text)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{str}{string to print/log}
}
