% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{linspace}
\alias{linspace}
\title{Implementation of the \code{linspace} Function from Matlab}
\usage{
linspace(a, b, n = 100)
}
\arguments{
\item{a}{initial number}

\item{b}{final number}

\item{n}{number of elements  (integer >= 2)}
}
\value{
vector of numbers from \code{a} to \code{b} with
\code{n} linearly spaced apart
}
\description{
Creates a vector of n elements equally spaced apart.
}
\examples{
linspace(0,100, 20)
}
