% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{logspace}
\alias{logspace}
\title{Implementation of the \code{logspace} Function from Matlab}
\usage{
logspace(a, b, n = 100)
}
\arguments{
\item{a}{initial number}

\item{b}{final number}

\item{n}{number of elements  (integer >=2)}
}
\value{
vector of numbers from \code{a} to \code{b} with
\code{n} logarithmically (base 10) spaced apart
}
\description{
Creates a vector of n elements logarithmically spaced apart.
}
\examples{
logspace(-2, 3,20)
}
