% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{make_forcing_function}
\alias{make_forcing_function}
\title{Makes Forcing Function From Times and Values}
\usage{
make_forcing_function(times, values, type, output_times)
}
\arguments{
\item{times}{time values for the forcing function}

\item{values}{magnitude for each time (same length of time)}

\item{type}{string indicating the type of forcing function can be one of the following:
\itemize{
      \item  \code{"step"} for constant values that switch to new values at the times
      \item  \code{"linear"} to linearly interpolate between the points
       }}

\item{output_times}{vector of simulation output times}
}
\value{
matrix with two columns: first column is a vector of times and the second column is a vector of values
}
\description{
Takes a list of times, values, and an interpolation method
}
\keyword{internal}
