% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{multiplot}
\alias{multiplot}
\title{Collect Multiple Plots}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{list of plot objects}

\item{plotlist}{list of plot objects}

\item{cols}{number of columns}
}
\value{
multiplot object
}
\description{
:
Adapted from here:
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
