% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{pad_string}
\alias{pad_string}
\title{Pad String with Spaces}
\usage{
pad_string(str, maxlength = 1, location = "beginning")
}
\arguments{
\item{str}{string}

\item{maxlength}{length to pad to}

\item{location}{either \code{"beginning"} to pad the left or \code{"end"} to pad the right}
}
\value{
Padded string
}
\description{
Adds spaces to the beginning or end of strings until it reaches the maxlength. Used for aligning text.
}
\examples{
pad_string("bob", maxlength=10)
pad_string("bob", maxlength=10, location="end")
}
