% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{parse_patterns}
\alias{parse_patterns}
\title{Parse String for Prototype Functions}
\usage{
parse_patterns(cfg, str)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{str}{string}
}
\value{
String with the prototype functions replaced
}
\description{
A string can contain any number of prototype functions, and this function will find them and replace them with the actual R code.
}
\keyword{internal}
