% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{prepare_figure}
\alias{prepare_figure}
\title{Make ggplot Figure Pretty}
\usage{
prepare_figure(purpose, fo, y_tick_minor = FALSE, y_tick_major = FALSE,
  x_tick_minor = FALSE, x_tick_major = FALSE)
}
\arguments{
\item{purpose}{either \code{"present"}, \code{"print"} or \code{"shiny"}}

\item{fo}{ggplot figure object}

\item{y_tick_minor}{Boolean value to control grid lines}

\item{y_tick_major}{Boolean value to control grid lines}

\item{x_tick_minor}{Boolean value to control grid lines}

\item{x_tick_major}{Boolean value to control grid lines}
}
\value{
ggplot object
}
\description{
Takes a ggplot object and alters the line thicknesses and makes
other cosmetic changes to make it more appropriate for exporting.
}
