% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{run_simulation_titrate}
\alias{run_simulation_titrate}
\title{Simulate With Titration or Rule-Based Inputs}
\usage{
run_simulation_titrate(SIMINT_p, SIMINT_cfg)
}
\arguments{
\item{SIMINT_p}{list of system parameters}

\item{SIMINT_cfg}{ubiquity system object}
}
\value{
som
}
\seealso{
\code{\link{system_new_tt_rule}}, \code{\link{system_set_tt_cond}} and the titration vignette (\code{vignette("Titration", package = "ubiquity")})
}
