% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{run_simulation_ubiquity}
\alias{run_simulation_ubiquity}
\title{Simulate Individual Response}
\usage{
run_simulation_ubiquity(SIMINT_parameters, SIMINT_cfg,
  SIMINT_dropfirst = TRUE)
}
\arguments{
\item{SIMINT_parameters}{vector of parameters}

\item{SIMINT_cfg}{ubiquity system object}

\item{SIMINT_dropfirst}{when \code{TRUE} it will drop the first sample point (prevents bolus doses from starting at 0)}
}
\value{
The simulation output is mapped (\code{som}) is a list.
time-course is stored in the \code{simout} element. The first column (\code{time})
contains the simulation time in the units of the simulation, days in this
case. Next there is a colunn for each state and a
column for each output. Each system parameter is also
passed through the simulation into the output. Models with covariates
will contain the  initial value (prefix: \code{SIMINT_CVIC_}) as well as the values at each time point. For the
Next secondary parameters
are also provided as well as a column for each timescale specified in
the system file with a "\code{ts.}" prefix.
}
\seealso{
Simulation vignette (\code{vignette("Simulation", package = "ubiquity")})
}
