% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{sample_around}
\alias{sample_around}
\title{Define Sample Times Around Events}
\usage{
sample_around(tvals, ot, sample_delta_mult = 1e-06)
}
\arguments{
\item{tvals}{vector of event times}

\item{ot}{simualtion output times}

\item{sample_delta_mult}{multiplier used to control the magnitude of spacing around and following event times}
}
\value{
vector of event times and added samples
}
\description{
When events, such as bolus doses, are applied to the system
rapid changes can occur. If the system is not sampled heavily around these
times, these changes may be missed in the output profiles. Based on the total
duration of the sample times, extra samples can be added near these events.
}
\details{
For more information on setting options for population simulation see the
stochastic section of the \code{\link{system_set_option}} help file.
}
\keyword{internal}
