% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{simulate_subjects}
\alias{simulate_subjects}
\title{Run Population Simulations}
\usage{
simulate_subjects(parameters, cfg, progress_message = "Simulating Subjects:")
}
\arguments{
\item{parameters}{list containing the typical value of parameters}

\item{cfg}{ubiquity system object}
}
\value{
Mapped simulation output with individual predictions, individual
parameters, and summary statistics of the parameters. The Vignettes below
details on the format of the output.
}
\seealso{
Vignette on simulation (\code{vignette("Simulation", package = "ubiquity")}) titration (\code{vignette("Titration", package = "ubiquity")}) as well as \code{\link{som_to_df}}
}
