% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{som_to_df}
\alias{som_to_df}
\title{Converts the Wide/Verbose Output Simulation Functions into Data Frames}
\usage{
som_to_df(cfg, som)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{som}{simulation output from \code{run_simulation_ubiquity}, \code{simulate_subjects},  \code{run_simulation_titrate}}
}
\value{
r1
}
\seealso{
\code{\link{run_simulation_titrate}} internally when running simulations.
}
