% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_check_requirements}
\alias{system_check_requirements}
\title{Check For Perl and C Tools}
\usage{
system_check_requirements(checklist = list(perl = list(check = TRUE,
  perlcmd = "perl"), C = list(check = TRUE)), verbose = TRUE)
}
\arguments{
\item{checklist}{list with names corresponding to elements of the system to check.}

\item{verbose}{enable verbose messaging}
}
\value{
List fn result of all packages
}
\description{
Check the local installation for perl and verify C compiler is installed and working.
}
\examples{
\donttest{
invisible(system_check_requirements())
}
}
