% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_clear_cohorts}
\alias{system_clear_cohorts}
\title{Clear all Cohorts}
\usage{
system_clear_cohorts(cfg)
}
\arguments{
\item{cfg}{ubiquity system object}
}
\value{
ubiquity system object with no cohorts defined
}
\description{
Clear previously defined cohorts
}
