% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_define_cohort}
\alias{system_define_cohort}
\title{Define Estimation Cohort}
\usage{
system_define_cohort(cfg, cohort)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{cohort}{list with cohort information}
}
\value{
ubiquity system object with cohort defined
}
\description{
Define a cohort to include in a parameter estimation
}
