% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_fetch_guess}
\alias{system_fetch_guess}
\title{Fetch Current Parameter Guesses}
\usage{
system_fetch_guess(cfg)
}
\arguments{
\item{cfg}{ubiquity system object}
}
\value{
list of current parameter gauesses
}
\description{
Fetch a list of the guesses for the current parameter set and
parameters selected for estimation
}
