% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_fetch_nca}
\alias{system_fetch_nca}
\title{Fetch NCA Results}
\usage{
system_fetch_nca(cfg, analysis_name = "analysis")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{analysis_name}{string containing the name of the NCA analysis (default \code{'analysis'})}
}
\value{
List with a data frame of the NCA results (\code{NCA_summary}), the raw
output from PKNCA (\code{PKNCA_results}), and also a list element indicating the
overall success of the function call (\code{isgood})
}
\description{
Fetches the NCA summary from the ubiquity system object.
}
\seealso{
Vignette on NCA (\code{vignette("NCA", package = "ubiquity")})
}
