% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_fetch_rpt_officer_object}
\alias{system_fetch_rpt_officer_object}
\title{Extracts the officer Object From the Specified ubiquity Report}
\usage{
system_fetch_rpt_officer_object(cfg, rptname = "default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{ubiquity report name}
}
\value{
officer report object
}
\description{
This will extract an officer object from the ubiqiuty system
object for the specified report name.
}
\seealso{
\code{\link{system_set_rpt_officer_object}}
}
