% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_fetch_rpt_onbrand_object}
\alias{system_fetch_rpt_onbrand_object}
\title{Extracts the onbrand Object From the Specified ubiquity Report}
\usage{
system_fetch_rpt_onbrand_object(cfg, rptname = "default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{ubiquity report name}
}
\value{
onbrand report object
}
\description{
This will extract an onbrand object from the ubiqiuty system
object for the specified report name.
}
\seealso{
\code{\link{system_set_rpt_onbrand_object}}
}
