% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_fetch_TSsys}
\alias{system_fetch_TSsys}
\title{Fetch System Timescale}
\usage{
system_fetch_TSsys(cfg)
}
\arguments{
\item{cfg}{ubiquity system object}
}
\value{
Name of the system timescale or \code{NULL} if it was not found
}
\description{
Reads through the system information and tries to determine the
system time scale (the timescale that has a value of 1)
}
