% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_glp_init}
\alias{system_glp_init}
\title{Initialize GLP study design}
\usage{
system_glp_init(cfg, study_title = "Study Title",
  study_name = "default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{study_title}{String containing descriptive information about the study}

\item{study_name}{short name used to identify the study in other functions  (\code{"default"})}
}
\value{
cfg ubiquity system object with the study initialized
}
\description{
Creates a new GLP study design
}
