% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_log_entry}
\alias{system_log_entry}
\title{Add Log Entry}
\usage{
system_log_entry(cfg, entry)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{entry}{string containing the log entry}
}
\description{
Appends a specified line to the analysis log
}
\keyword{internal}
