% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_nca_report}
\alias{system_nca_report}
\title{Add NCA output to ubiquity report}
\usage{
system_nca_report(cfg, ncares = NULL, maxrow = 15, label_data = FALSE,
  rptname = "default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{ncares}{result of (\code{\link{system_nca_run}})}

\item{maxrow}{maximum number of rows per slide for tables}

\item{label_data}{Boolean variable to control labeling of data on plots (\code{FALSE})}

\item{rptname}{optional name of the report (\code{"default"})}
}
\description{
Performs NCA in an automated fashion subsetting the data using muleiple criteria
}
\details{

}
