% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_nca_run}
\alias{system_nca_run}
\title{Automatic NCA}
\usage{
system_nca_run(cfg, DSNAME = "PKDS", dscale = 1, ssby = list(id = c(1, 2,
  3), dose_number = c(1, 2)), obs = list(time = "TIME", conc = "DV", dose =
  "AMT", id = "ID", dsinc = NULL))
}
\arguments{
\item{cfg}{ubiquity system object}

\item{DSNAME}{name of dataset loaded with (\code{\link{system_load_data}})}

\item{dscale}{factor to multiply the dose to get it into the same units as concentration:
if you are dosing in mg/kg and your concentrations is in ng/ml, then \code{dscale = 1e6}}

\item{ssby}{list specifying how to subset the data, with list names being
column names from the dataset and values being those used for subsetting (note
every combination will be considered when performing NCA:
        If your dataset specifies the dose number (\code{dose_number}) for subjects and you want to perform NCA on doses 1 and 5 for subjects (\code{id}) 2, 8 and 120 then ssby would look like:
\preformatted{
 ssby = list(id=c(2,8,120),  dose_number=c(1,5))
}}

\item{obs}{list with names specifying the time, concentration, dose, and id columns to use when performing NCA}

\item{dsinc}{character vector of columns from the dataset to include in the output summary: These should be constant for the combinations specified in \code{ssby}}
}
\description{
Performs NCA in an automated fashion subsetting the data using muleiple criteria
}
