% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_new}
\alias{system_new}
\title{Create New \code{system.txt} File}
\usage{
system_new(file_name = "system.txt", system_file = "template",
  overwrite = FALSE)
}
\arguments{
\item{file_name}{name of the new file to create}

\item{system_file}{name of the system file to copy}

\item{overwrite}{if \code{TRUE} the new system file will overwrite any existing files present}
}
\value{
\code{TRUE} if the new file was created and \code{FALSE} otherwise
}
\description{
Copy a blank template (\code{system_file="template"}) file to the working directory or an example by specifying the following:

\itemize{
  \item \code{"template"} - Empty system file template
  \item \code{"mab_pk"} - General compartmental model of mAb PK from Davda 2014 http://doi.org/10.4161/mabs.29095
  \item \code{"pbpk"} - PBPK model of mAb disposition in mice from Shah 2012 
}
}
\examples{
# To create an empty template:
system_new("system.txt", TRUE)
}
