% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_new_list}
\alias{system_new_list}
\title{Fetch List of Available System Templates}
\usage{
system_new_list()
}
\value{
list with the template names as the keys
\itemize{
\item{file_path} Full path to the system file
\item{description} Description of what this system file provides
}
}
\description{
Returns a list of internal templates with descriptions of their contents and file locations
}
\examples{
# To get a list of systems
systems = system_new_list()
}
