% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_nm_check_ds}
\alias{system_nm_check_ds}
\title{title
Description}
\usage{
system_nm_check_ds(cfg, DS = "DSNAME", col_ID = "ID", col_CMT = "CMT",
  col_DV = "DV", col_TIME = "TIME", col_AMT = "AMT", col_RATE = "RATE",
  col_EVID = "EVID", col_GROUP = NULL, filter = NULL, INPUTS = NULL,
  OBS = NULL)
}
\arguments{
\item{p1}{}
}
\value{
r1
}
\examples{
# Examples
}
