% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_doc_format_section}
\alias{system_report_doc_format_section}
\title{Formats the Current Document Section}
\usage{
system_report_doc_format_section(
  cfg,
  rptname = "default",
  section_type = NULL,
  w = NULL,
  h = NULL,
  sep = FALSE,
  widths = NULL,
  space = NULL
)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{report name initialized with \code{system_report_init}}

\item{section_type}{type of section to apply, either \code{"columns"},  \code{"continuous"}, \code{"landscape"}, \code{"portrait"}, \code{"columns"},  or  \code{"columns_landscape"}}

\item{w}{width  in inches of the section of the page (\code{NULL})}

\item{h}{height in inches of the section of the page (\code{NULL})}

\item{sep}{Boolean value controlling line separating columns (\code{FALSE})}

\item{widths}{column widths in inches, number of columns set by number of values (\code{NULL})}

\item{space}{space in inches between columns (\code{NULL})}
}
\value{
cfg ubiquity system object with the section added to the body
}
\description{
Sets the section format for the content added since the last section content type was defined.  This is a wrapper for the body_end_section* functions in officer.
}
