% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_doc_set_ph}
\alias{system_report_doc_set_ph}
\title{Sets Placeholder Content for Word Document Report}
\usage{
system_report_doc_set_ph(cfg, rptname = "default", phname = NULL,
  phcontent = NULL, phlocation = "body")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{report name initialized with \code{system_report_init}}

\item{phname}{name of the placeholder}

\item{phcontent}{content to be replaced}

\item{phlocation}{location of the placeholder: \code{"body"} (default), \code{"header"}, or \code{"footer"}}
}
\value{
cfg ubiquity system object with the placeholder content set
}
\description{
Adds or updates content to be substituted for placeholders in the specified report.  

For example if you have <HEADER_LEFT> in the header of your document and you wanted to
replace it with the text "Upper left" you would do the following:

\code{
  cfg = system_report_doc_set_ph(cfg, 
        phcontent  = "Upper Left" ,
        phname     = "HEADER_LEFT", 
        phlocation = "header")}

Notice the \code{phname} just has \code{HEADER_LEFT} and leaves off the \code{<>}
}
