% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_estimation}
\alias{system_report_estimation}
\title{Generate a Report from Parameter Estimation}
\usage{
system_report_estimation(cfg, rptname = "default", analysis_name = NULL)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{report name (either PowerPoint or Word)}

\item{analysis_name}{string containing the name of the estimation analysis and used as a prefix to store the results}
}
\value{
ubiquity system object with estimation report appended
}
\description{
This will take the output generated during a parameter estimation and append those results to a specified report.
}
\seealso{
\code{\link{system_report_init}}, the reporting vignette (\code{vignette("Reporting", package = "ubiquity")})
and the estimation vignette (\code{vignette("Estimation", package = "ubiquity")})
}
