% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_glp}
\alias{system_report_glp}
\title{Report GLP Study}
\usage{
system_report_glp(
  cfg,
  study_title = "Study Title",
  study_name = "default",
  rptname = "default"
)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{study_title}{String containing descriptive information about the study}

\item{study_name}{short name used to identify the study in other functions  (\code{"default"})}

\item{rptname}{short name used to identify the report to attach results to the study in other functions (\code{"default"})}
}
\value{
cfg ubiquity system object with the study report information added
}
\description{
Append GLP study design a report
}
