% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_init}
\alias{system_report_init}
\title{Initialize a New officer Report}
\usage{
system_report_init(cfg, template = NULL, rptname = "default",
  meta = NULL)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{template}{path to template file (\code{NULL} will load the default ubiquity template)}

\item{rptname}{report name}

\item{meta}{list containing metadata identifying relevant indices for slide layouts}
}
\description{
Creates a new officer report based either on the ubiquity
template or one specified by the user. Once created, slides can then be
added.
}
\seealso{
Reporting vignette (\code{vignette("Reporting", package = "ubiquity")})
}
