% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_set}
\alias{system_report_set}
\title{Overwrite officer Object for a Given Report}
\usage{
system_report_set(cfg, rptname = "default", rpt = NULL)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{report name initialized with \code{system_report_init}}

\item{rpt}{officer object}
}
\value{
ubiquity system object with \code{rpt} as content for \code{rptname}
}
\description{
Replace the report named \code{rptname} with the contents in \code{rpt}
}
\seealso{
\code{\link{system_report_init}} and \code{\link{system_fetch_report}}
}
