% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_slide_content}
\alias{system_report_slide_content}
\title{Add Slide With Main Body of Content}
\usage{
system_report_slide_content(
  cfg,
  title = "Title",
  sub_title = NULL,
  rptname = "default",
  content_type = "text",
  content = "Text"
)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{title}{string with slide title (\code{"Title"})}

\item{sub_title}{string with slide sub title (code{NULL})}

\item{rptname}{report name initialized with \code{system_report_init}}

\item{content_type}{type of content for main body of slide}

\item{content}{content of main body of slide}
}
\value{
ubiquity system object with slide added to report
}
\description{
Creates a report slide with a title and single large area of content
}
\details{
For information on the format of content, see \code{\link{system_report_ph_content}}.
}
\seealso{
\code{\link{system_report_init}} and the reporting vignette (\code{vignette("Reporting", package = "ubiquity")})
}
