% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_slide_section}
\alias{system_report_slide_section}
\title{Generate Slide with Section Break}
\usage{
system_report_slide_section(cfg, title = "Title", sub_title = NULL,
  rptname = "default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{title}{string with slide title}

\item{sub_title}{string with slide sub title}

\item{rptname}{report name initialized with \code{system_report_init}}
}
\value{
ubiquity system object with slide added to report
}
\seealso{
\code{\link{system_report_init}} and the reporting vignette (\code{vignette("Reporting", package = "ubiquity")})
}
