% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_slide_title}
\alias{system_report_slide_title}
\title{Generate Title Slide}
\usage{
system_report_slide_title(cfg, title = "Title", sub_title = NULL,
  rptname = "default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{title}{string with slide title (\code{"Title"})}

\item{sub_title}{string with slide sub title (code{NULL})}

\item{rptname}{report name initialized with \code{system_report_init}}
}
\value{
ubiquity system object with slide added to report
}
\description{
Creates a report title slide.
}
\seealso{
\code{\link{system_report_init}} and the reporting vignette (\code{vignette("Reporting", package = "ubiquity")})
}
