% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_rpt_add_doc_content}
\alias{system_rpt_add_doc_content}
\title{Adds Content to a Word Report}
\usage{
system_rpt_add_doc_content(
  cfg,
  type = NULL,
  content = NULL,
  rptname = "default"
)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{type}{Type of content to add. See the
onbrand function \code{\link[onbrand]{report_add_doc_content}} 
for the allowed content types.}

\item{content}{List with content to add to the report.  See the
onbrand function \code{\link[onbrand]{report_add_doc_content}} 
format of this list.}

\item{rptname}{Report name}
}
\value{
ubiquity system object with the content added to the specified
report
}
\description{
Appends content to an open ubiquity Word report.
}
\seealso{
\code{\link[onbrand]{report_add_doc_content}} and
Reporting vignette (\code{vignette("Reporting", package = "ubiquity")})
}
