% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_rpt_nca}
\alias{system_rpt_nca}
\title{Report NCA}
\usage{
system_rpt_nca(
  cfg,
  rptname = "default",
  analysis_name = "analysis",
  rows_max = 10,
  table_headers = TRUE
)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{report name (either PowerPoint or Word)}

\item{analysis_name}{string containing the name of the NCA analysis (default \code{'analysis'})}

\item{rows_max}{maximum number of rows per slide when generating tabular data}

\item{table_headers}{Boolean variable to add descriptive headers to output tables (default \code{TRUE})}
}
\value{
cfg ubiquity system object with the NCA results appended to the specified report and if the analysis name is specified:
}
\description{
Appends the results of NCA to a report
}
\seealso{
Vignette on NCA (\code{vignette("NCA", package = "ubiquity")})
}
