% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_rpt_save_report}
\alias{system_rpt_save_report}
\title{Save Report to a File}
\usage{
system_rpt_save_report(cfg, output_file = NULL, rptname = "default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{output_file}{File to save the report to (must be either .pptx or .docx
depending on the type of report)}

\item{rptname}{ubiquity report name}
}
\value{
list with the follwoing elements 
\itemize{
\item{isgood} Boolean variable indicating success or failure
\item{msgs}   Verbose description of the save results
}
}
\description{
Saves a ubiquity report to the specified file.
}
\seealso{
Reporting vignette (\code{vignette("Reporting", package = "ubiquity")})
}
