% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_rpt_template_details}
\alias{system_rpt_template_details}
\title{Generate Details about Report Template}
\usage{
system_rpt_template_details(cfg, rptname = "default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{Report name}
}
\value{
list with template information, see
\code{\link[onbrand]{template_details}} for information on the structure of
this list.
}
\description{
Wrapper for the onbrand::template_details function, see the
help for that function for more information
}
\seealso{
\code{\link[onbrand]{template_details}} and
Reporting vignette (\code{vignette("Reporting", package = "ubiquity")})
}
