% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_set_bolus}
\alias{system_set_bolus}
\title{Set Bolus Inputs}
\usage{
system_set_bolus(cfg, state, times, values)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{state}{name of the state to apply the bolus}

\item{times}{list of injection times}

\item{values}{corresponding list injection values}
}
\value{
Ubiquity system object with the bolus information set
}
\description{
Defines infusion rates specified in the system file using  \code{<B:times>} and   \code{<B:events>}
}
\examples{
# cfg = system_set_bolus(cfg,
#            state  = "SNAME",
#            times  = c(0, 1), 
#            values = c(10, 0))
# Examples
}
\seealso{
\code{\link{system_zero_inputs}}
}
