% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_set_covariate}
\alias{system_set_covariate}
\title{Set Covariate Values}
\usage{
system_set_covariate(cfg, covariate, times, values)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{covariate}{name of the covariate}

\item{times}{list of times (system time units)}

\item{values}{corresponding list of values}
}
\value{
Ubiquity system object with the covariate set
}
\description{
Covariates specified in the system file using  \code{<CV:?>}
and \code{<CVSET:?:?>} will have their default values for a given parameter
set. This function is a means to overwrite those values.
}
\examples{
#  cfg = system_set_covariate(cfg, 
#                             covariate = "COV",
#                             times     = c(1, 3, 25),
#                             values    = c(1, 2,  1))
}
