% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_set_iiv}
\alias{system_set_iiv}
\title{Set Variability Terms}
\usage{
system_set_iiv(cfg, IIV1, IIV2, VALUE)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{IIV1}{row name of the variance/covariance matrix}

\item{IIV2}{column name of the variance/covariance matrix element}

\item{VALUE}{value to assign to the variance/covariance matrix element}
}
\value{
Ubiquity system object with IIV information set
}
\description{
Set elements of the current variance covariance matrix
specified in the system file with \code{<IIV:?:?> ?}, \code{<IIVCOR:?:?>?}, \code{<IIVSET:?:?> ?}, \code{<IIVCORSET:?:?>?}
}
\examples{
# cfg = system_set_iiv(cfg,
#                      IIV1 = "ETACL",
#                      IIV2 = "ETAVc",
#                      VALUE=0.03)
}
\seealso{
\code{\link{system_fetch_iiv}}
}
