% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_set_option}
\alias{system_set_option}
\title{Setting Analysis Options}
\usage{
system_set_option(cfg, group, option, value)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{group}{options are grouped together by the underlying activity being performed: "solver", "stochastic", "simulation", "estimation", "logging", or "titration"}

\item{option}{for each group there are a set of options}

\item{value}{corresponding value for the option}
}
\value{
Ubiquity system object with the option set
}
\description{
Different options associated performing analyses (e.g running
simulations, performing parameter esitmations, logging, etc.) can be set
with this function
}
