% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_set_parameter}
\alias{system_set_parameter}
\title{Set Value for Parameter}
\usage{
system_set_parameter(cfg, parameters, pname, value)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{parameters}{vector of parameters}

\item{pname}{parameter name}

\item{value}{value}
}
\value{
parameters vector with \code{pname} set to \code{value}
}
\description{
Assigns a value for a named parameter in a parameter list.
}
\details{
To set the parameter Vc to a value of 3, the following would be used:
\preformatted{parameters = system_fetch_parameters(cfg) 
parameters = system_set_parameter(cfg, parameters, pname = 'Vc', value = 3) 
}
}
