% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_set_rate}
\alias{system_set_rate}
\title{Set Infusion Rate Inputs}
\usage{
system_set_rate(cfg, rate, times, levels)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rate}{name of infusion rate}

\item{times}{list of time values}

\item{levels}{corresponding list of infusion values}
}
\value{
Ubiquity system object with the infusion rate set
}
\description{
Defines infusion rates specified in the system file using  \code{<R:?>}
}
\examples{
# cfg = system_set_rate(cfg,
#            rate   = "RNAME",
#            times  = c(0, 1), 
#            levels = c(10, 0))
# Examples
}
\seealso{
\code{\link{system_zero_inputs}}
}
