% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_set_rpt_officer_object}
\alias{system_set_rpt_officer_object}
\title{Sets the officer Object for the Specified ubiquity Report}
\usage{
system_set_rpt_officer_object(cfg, rpt = NULL, rptname = "default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rpt}{officer report object}

\item{rptname}{ubiquity report name}
}
\value{
ubiquity system object with the replaced officer object
}
\description{
This will replace the officer object in the ubiqiuty system
object for the specified report name with the value supplied.
}
\seealso{
\code{\link{system_fetch_rpt_officer_object}}
}
