% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_set_rpt_onbrand_object}
\alias{system_set_rpt_onbrand_object}
\title{Sets the onbrand Object for the Specified ubiquity Report}
\usage{
system_set_rpt_onbrand_object(cfg, obnd = NULL, rptname = "default")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{obnd}{onbrand report object}

\item{rptname}{ubiquity report name}
}
\value{
ubiquity system object with onbrand report set
}
\description{
This will reset the onbrand object in the ubiqiuty system
object for the specified report name.
}
\seealso{
\code{\link{system_fetch_rpt_onbrand_object}}
}
