% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_simulate_estimation_results}
\alias{system_simulate_estimation_results}
\title{Simulate Results at Estimates}
\usage{
system_simulate_estimation_results(pest, cfg)
}
\arguments{
\item{pest}{vector of parameters}

\item{cfg}{ubiquity system object}
}
\value{
observations in a list, see \code{\link{system_define_cohort}} when \code{estimation=FALSE}
}
\description{
Simulates the system at the parameter estimates \code{pest}
}
\seealso{
\code{\link{system_define_cohort}} \code{\link{system_plot_cohorts}}
}
