% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_ts_to_simtime}
\alias{system_ts_to_simtime}
\title{Convert Time in Timescale to Simulation Time}
\usage{
system_ts_to_simtime(cfg, tstime, ts)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{tstime}{numeric time of the timescale}

\item{ts}{string containing the timescale}
}
\value{
\code{tstime} in the system timescale units
}
\description{
converts a time specified in a defined timescale (say weeks) to the
timescale of the simulation (say hours if the rates are in 1/hr units)
}
\keyword{internal}
