% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{tic}
\alias{tic}
\title{Implementation of Matlab \code{tic()} command}
\usage{
tic(gcFirst = TRUE, type = c("elapsed", "user.self", "sys.self"))
}
\arguments{
\item{gcFirst}{controls garbage collection}

\item{type}{can be either \code{"elapsed"} \code{"user.self"} or \code{"sys.self"}}
}
\value{
time tic was called
}
\description{
Used in conjunction with \code{toc()} to find the elapsed time
when code is executed. Adapted from:
http://stackoverflow.com/questions/1716012/stopwatch-function-in-r
}
\examples{
tic()
Sys.sleep(3)
toc()
}
\seealso{
\code{\link{toc}}
}
