% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{ubiquity_name_check}
\alias{ubiquity_name_check}
\title{Check Names of Cohorts, Analyses, Reports, etc.}
\usage{
ubiquity_name_check(test_name)
}
\arguments{
\item{test_name}{string containing the name to be tested}
}
\value{
List with Boolean element \code{isgood} that is \code{TRUE} when the name tests correct, \code{FALSE} when it fails. The element \code{msgs} contains a verbose message on why it fails.
}
\description{
Checks names specified for different analysis aspects (cohorts,
analyses, reports, etc.) to make sure that they start with a letter and
contain only letters, numbers and _
}
\keyword{internal}
