% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{var2string_gen}
\alias{var2string_gen}
\title{General Method to Convert a Variable into a String for Automated Reporting Purposes}
\usage{
var2string_gen(var)
}
\arguments{
\item{var}{R variable}
}
\value{
Variable in string form
}
\examples{
# var2string_gen(c(1,2,3))
}
\keyword{internal}
