% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{vp}
\alias{vp}
\title{Print and Log Messages}
\usage{
vp(cfg, str)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{str}{sequence of strings to print}
}
\description{
Used to print messages to the screen and the log file.
}
\examples{
\donttest{
# Creating a system file from the mab_pk example
system_new(system_file="mab_pk", file_name="system_example.txt", overwrite=TRUE)

# Building the system 
cfg = build_system("system_example.txt")

# Initialzing the log file
vp(cfg, "Message that will be logged")
}
}
