% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{workshop_fetch}
\alias{workshop_fetch}
\title{Fetch Ubiquity Workshop Sections}
\usage{
workshop_fetch(section = "Simulation", overwrite = FALSE)
}
\arguments{
\item{section}{Name of the section of workshop to retrieve}

\item{overwrite}{if \code{TRUE} the new system file will overwrite any existing files present}

\item{on}{Name of the section of workshop to retrieve}
}
\value{
list
}
\description{
With the ubiquity package this function can be used to fetch
sections of the workshop.
}
\details{
Valid sections are "Simulation", "Estimation", "Titration" and "Reporting"
}
