% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{check_steady_state}
\alias{check_steady_state}
\title{Verify System Steady State}
\usage{
check_steady_state(cfg, som)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{som}{output of \code{\link{run_simulation_ubiquity}}}
}
\value{
list with name \code{steady_state} (boolean indicating weather the system was at steady state) and \code{states} a vector of states that have steady state offset.
}
\description{
Takes the output  \code{\link{run_simulation_ubiquity}} and verifies that the system is running at steady state by analyzing the timecourse of all of the states in the system
}
\keyword{internal}
