% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Performs parameter estimation}
\usage{
estimate_parameters(cfg)
}
\arguments{
\item{cfg}{ubiquity system object}
}
\value{
list with elements: \code{estimate} - vector of parameter estimates,
\code{raw} - raw output from the underlying optimization routine, 
\code{obj} - objective function value,  \code{statistics_est} - solution
statistics
}
\description{
Performs the actual parameter estimation
}
\keyword{internal}
