% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Performs parameter estimation}
\usage{
estimate_parameters(cfg)
}
\arguments{
\item{cfg}{ubiquity system object}
}
\value{
list with elements: 
\itemize{
\item \code{estimate} - vector of parameter estimates
\item \code{raw} - raw output from the underlying optimization routine
\item \code{conv} - list of convergence criteria with keys \code{num} (numeric)  \code{text} (text description)
\item \code{obj} - objective function value
\item \code{statistics_est} - solution statistics
\item \code{sysup} - Text to update the system file with the parameter estimates
\item \code{estimate} - Names list of parameter estiamtes
\item \code{report} - Named list with elements for reporting
\item \code{cohorts} - Cohort elements from the \code{cfg} file
\item \code{cohort_view} - Cohort view from \code{system_view()} 
\item \code{system_file} - Contents of the system file when estimation was
run
}
}
\description{
Performs the actual parameter estimation
}
\keyword{internal}
