% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{gg_log10_yaxis}
\alias{gg_log10_yaxis}
\title{Make Pretty ggplot y-Axis Log 10 Scale}
\usage{
gg_log10_yaxis(fo, ylim_min = NULL, ylim_max = NULL,
  y_tick_label = TRUE, x_tick_label = TRUE)
}
\arguments{
\item{fo}{ggplot figure object}

\item{ylim_min}{set to a number to define the lower bound of the y-axis}

\item{ylim_max}{set to a number to define the upper bound of the y-axis}

\item{y_tick_label}{\code{TRUE} to show y tick labels, \code{FALSE} to hide the y tick labels}

\item{x_tick_label}{\code{TRUE} to show x tick labels, \code{FALSE} to hide the x tick labels}
}
\value{
ggplot object with formatted axis
}
\description{
Wrapper for \code{\link{gg_axis}} to create a log 10 y-axis
}
\examples{
library("ggplot2")
df = data.frame(x = seq(0.01,10,.01),
               y = seq(0.01,10,.01)^2)
p       = ggplot(df, aes(x=x, y=y)) + geom_line()
# pretty up the axes
p       = prepare_figure(fo=p, purpose="print")
# pretty log10 y-axis 
p_logy  = gg_log10_yaxis(fo=p)
# pretty log10 x-axis 
p_logx  = gg_log10_xaxis(fo=p)
# pretty log10 yx-axis 
p_logxy = gg_axis(fo=p)
}
\seealso{
\code{\link{gg_axis}} and \code{\link{gg_log10_xaxis}}
}
