% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{GUI_log_entry}
\alias{GUI_log_entry}
\title{Wrapper for system_log_entry Used in ShinyApp}
\usage{
GUI_log_entry(cfg, text)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{text}{string to print/log}
}
\value{
Boolean variable indicating success (\code{TRUE}) or failure (\code{FALSE})
}
\description{
Called from the ShinyApp to add a log entry with "App"
prepended to the log entry
}
\keyword{internal}
