% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_fetch_parameters}
\alias{system_fetch_parameters}
\title{Get System Parameters}
\usage{
system_fetch_parameters(cfg)
}
\arguments{
\item{cfg}{ubiquity system object}
}
\value{
List of parameters for the selected parameter set or the default parameter set if the selected set does not exist
}
\description{
Fetch the parameters of the currently selected parameter set. To switch
between parameter sets use \code{\link{system_select_set}}
}
\examples{
# parameters = system_fetch_parameters(cfg)
}
