% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_load_data}
\alias{system_load_data}
\title{Loading Datasets}
\usage{
system_load_data(cfg, dsname, data_file, data_sheet)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{dsname}{short name of the dataset to be used to link this dataset to different operations}

\item{data_file}{the file name of the dataset or a data frame containing the data}

\item{data_sheet}{argument identifying the name of the sheet in an excel file}
}
\value{
Ubiquity system object with the dataset loaded
}
\description{
Loads datasets at the scripting level from  a variable if
\code{data_file} is a data.frame or from the following
formats (based on the file extension)
\itemize{
\item csv - comma delimited 
\item tab - tab delimited
\item xls - excel spread sheet
}

Multiple datasets can be loaded as long as they are given different
names. Datasets should be in a NONMEM-ish format with the
first row containing the column header names.
}
