% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_log_entry}
\alias{system_log_entry}
\title{Add Log Entry}
\usage{
system_log_entry(cfg, entry)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{entry}{string containing the log entry}
}
\description{
Appends a specified line to the analysis log
}
\examples{
\donttest{
# Creating a system file from the mab_pk example
system_new(system_file="mab_pk", file_name="system_example.txt", overwrite=TRUE)

# Building the system 
cfg = build_system("system_example.txt")

# Initialzing the log file
system_log_entry(cfg, "Text of log entry")
}
}
\keyword{internal}
