% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_nca_run}
\alias{system_nca_run}
\title{Automatic NCA}
\usage{
system_nca_run(cfg, DSNAME = "PKDS", dscale = 1, NCA_min = 4,
  analysis_name = "analysis", rescorr = TRUE, dsfilter = NULL,
  dsmap = list(TIME = "TIME", NTIME = "NTIME", CONC = "DV", DOSE = "AMT",
  ID = "ID", DOSENUM = NULL), digits = 3, dsinc = NULL)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{DSNAME}{name of dataset loaded with (\code{\link{system_load_data}})}

\item{dscale}{factor to multiply the dose to get it into the same units as concentration (default \code{1}):
if you are dosing in mg/kg and your concentrations is in ng/ml, then \code{dscale = 1e6}}

\item{NCA_min}{minimum number of points required to perform NCA for a given subset (default \code{4})}

\item{analysis_name}{string containing the name of the analysis (default 'analysis') to archive to files and reference results later}

\item{rescorr}{Boolean variable to correct for residual drug in a multiple dose setting (default \code{TRUE}): If there is an observation before the first observation of a given subset, that concentration will subtracted from the values of the subset}

\item{dsfilter}{list of names corresponding to the column names in the dataset and values are a sequence indicating values to keep (default \code{NULL}. Multiple names are and-ed together. For example the following would keep all of the records where dose is 1, 2, or 5 and the dose_number is 1
\preformatted{
 dsfilter = list(dose=c(1,2,5), dose_number = c(1))
}}

\item{dsmap}{list with names specifying the time (TIME), nominal time since last dose (NTIME), concentration (CONC), dose (DOSE), and id (ID) columns to use when performing NCA}

\item{digits}{number of significant digits to report (\code{3}), set to \code{NULL} to disable rounding}

\item{dsinc}{character vector of columns from the dataset to include in the output summary.}
}
\value{
cfg ubiquity system object with the NCA results and if the analysis name is specified:
\itemize{
    \item{output/{analysis_name}-nca_summary.csv} NCA summary 
    \item{output/{analysis_name}-pknca_summary.csv} Raw output from PKNCA with subject and dose number columns appended 
    \item{output/{analysis_name}-nca_data.RData} objects containing the NCA summary and a list with the ggplot grobs
}
}
\description{
Performs NCA in an automated fashion
}
