% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_init}
\alias{system_report_init}
\title{Initialize a New officer Report}
\usage{
system_report_init(cfg, template = NULL, rptname = "default", meta = NULL)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{template}{path to template file (\code{NULL} will load the default ubiquity template)}

\item{rptname}{report name}

\item{meta}{list containing metadata identifying relevant indices for slide layouts}

\item{cfg}{ubiquity system object with the named report initialized}
}
\seealso{
Reporting vignette (\code{vignette("Reporting", package = "ubiquity")})
}
