% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_save}
\alias{system_report_save}
\title{Save Report to File}
\usage{
system_report_save(cfg, rptname = "default",
  output_file = "myreport.pptx")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{report name initialized with \code{system_report_init}}

\item{output_file}{name of file to place the report}
}
\description{
Save the contents of \code{rptname} to the file \code{output_file}
}
\seealso{
\code{\link{system_report_init}}
}
