% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_report_view_layout}
\alias{system_report_view_layout}
\title{Generate Annotated Layout for pptx Template}
\usage{
system_report_view_layout(cfg, rptname = "default",
  output_file = "layout.pptx")
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rptname}{report name initialized with \code{system_report_init}}

\item{output_file}{name of file to place the annotated layout information, set to \code{NULL} to suppress file generation}
}
\value{
officer pptx object with the layout of the template annotated,
}
\seealso{
\code{\link{system_report_init}} and the reporting vignette (\code{vignette("Reporting", package = "ubiquity")})
}
