% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_req}
\alias{system_req}
\title{Require Suggested Packages}
\usage{
system_req(pkgs)
}
\arguments{
\item{pkgs}{character vector of s}
}
\value{
Boolean result of all packages
}
\description{
Used to ensure packages are loaded as they are needed.
}
\keyword{internal}
