% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_req}
\alias{system_req}
\title{Require Suggested Packages}
\usage{
system_req(pkgs)
}
\arguments{
\item{pkgs}{character vector of package names}
}
\value{
Boolean result (one for each package) of the loaded status
}
\description{
Used to ensure packages are loaded as they are needed.
}
\keyword{internal}
