% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_req}
\alias{system_req}
\title{Require Suggested Packages}
\usage{
system_req(pkgs)
}
\arguments{
\item{pkgs}{character vector of package names to check}
}
\value{
Boolean result of the loaded (stand alone) or installed (package) status for all of the packages
}
\description{
Used to ensure packages are loaded as they are needed for the
stand alone distribution of ubiquity. If the ubiquity package is being used this
function simply returns 'TRUE' if the packages are installed and FALSE if
if not.
}
\keyword{internal}
