% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_set_rate}
\alias{system_set_rate}
\title{Set Infusion Rate Inputs}
\usage{
system_set_rate(cfg, rate, times, levels)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{rate}{name of infusion rate}

\item{times}{list of time values}

\item{levels}{corresponding list of infusion values}
}
\value{
Ubiquity system object with the infusion rate set
}
\description{
Defines infusion rates specified in the system file using  \code{<R:?>}
}
\examples{
\donttest{
# Creating a system file from the mab_pk example
system_new(system_file="mab_pk", file_name="system_example.txt", overwrite=TRUE)

# Building the system 
cfg = build_system("system_example.txt")

# 5 minute infusion at 10 mg/min
cfg = system_set_rate(cfg,
           rate   = "Dinf",
           times  = c(0,  5), 
           levels = c(10, 0))
}
}
\seealso{
\code{\link{system_zero_inputs}}
}
