% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{system_simulate_estimation_results}
\alias{system_simulate_estimation_results}
\title{Simulate Results at Estimates}
\usage{
system_simulate_estimation_results(pest, cfg, details = FALSE)
}
\arguments{
\item{pest}{vector of parameters}

\item{cfg}{ubiquity system object}

\item{details}{set \code{TRUE} to display information about cohorts as they are simulated (useful for debugging)}
}
\value{
observations in a list, see \code{\link{system_od_general}} when \code{estimation=FALSE}
}
\description{
Simulates the system at the parameter estimates \code{pest} for creating diagnostic plots
}
\seealso{
\code{\link{system_define_cohort}}, \code{\link{system_plot_cohorts}}
and the vignette on parameter estimation (\code{vignette("Estimation", package = "ubiquity")})
}
