% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{toc}
\alias{toc}
\title{Implementation of Matlab \code{toc()} command}
\usage{
toc()
}
\value{
time in seconds since tic() was called
}
\description{
Used in conjunction with \code{tic()} to find the elapsed time
when code is executed. Adapted from:
http://stackoverflow.com/questions/1716012/stopwatch-function-in-r
}
\examples{
tic()
Sys.sleep(3)
toc()
}
\seealso{
\code{\link{tic}}
}
