% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{var2string_gen}
\alias{var2string_gen}
\title{Convert R Objects to Strings}
\usage{
var2string_gen(var)
}
\arguments{
\item{var}{R variable}
}
\value{
Variable in string form
}
\description{
Mechanism for converting R objects strings for reporting.
}
\examples{
var2string_gen(c(1,2,3))
}
\keyword{internal}
