% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{vp}
\alias{vp}
\title{Print and Log Messages}
\usage{
vp(cfg, str)
}
\arguments{
\item{cfg}{ubiquity system object}

\item{str}{string to print}
}
\description{
Used to print messages to the screen and the log file.
}
