% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{workshop_fetch}
\alias{workshop_fetch}
\title{Fetch Sections of Ubiquity Workshop}
\usage{
workshop_fetch(section = "Simulation", overwrite = FALSE)
}
\arguments{
\item{section}{Name of the section of workshop to retrieve}
}
\value{
list
}
\details{
Valid sections are "Simulation", "Estimation", "Titration" and "Reporting"
}
