% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquity.R
\name{workshop_fetch}
\alias{workshop_fetch}
\title{Fetch Ubiquity Workshop Sections}
\usage{
workshop_fetch(section = "Simulation", overwrite = FALSE,
  output_directory = getwd())
}
\arguments{
\item{section}{Name of the section of workshop to retrieve  ("Simulation")}

\item{overwrite}{if \code{TRUE} the new system file will overwrite any existing files present}

\item{output_directory}{directory where workshop files will be placed (getwd())}
}
\value{
list
}
\description{
With the ubiquity package this function can be used to fetch
example files for different sections of the workshop.
}
\details{
Valid sections are "Simulation", "Estimation", "Titration" "Reporting", and "NCA"
}
\examples{
\donttest{
workshop_fetch("Estimation", output_directory=tempdir(), overwrite=TRUE)
}
}
